LiblineaR Classification Learner

LiblineaR Classification Learner

LiblineaR Classification Learner

Details

Type of SVC depends on type argument:

0 L2-regularized logistic regression (primal)

1 L2-regularized L2-loss support vector classification (dual)

3 L2-regularized L1-loss support vector classification (dual)

2 L2-regularized L2-loss support vector classification (primal)

4 Support vector classification by Crammer and Singer

5 L1-regularized L2-loss support vector classification

6 L1-regularized logistic regression

7 L2-regularized logistic regression (dual)

If number of records > number of features, type = 2 is faster than type = 1 (Hsu et al. 2003).

Note that probabilistic predictions are only available for types 0, 6, and 7. The default epsilon value depends on the type parameter, see [LiblineaR::LiblineaR].

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLiblineaR

Methods

Public methods

LearnerClassifLiblineaR$new()

LearnerClassifLiblineaR$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

#' Creates a new instance of this [R6][R6::R6Class] class.

Usage

LearnerClassifLiblineaR$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLiblineaR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.