Universal slot accessor function for siamcat-class.

This function is used internally by many accessors.

accessSlot(siamcat, slot, verbose=1)

Arguments

siamcat

an object of siamcat-class.

slot

A character string indicating the slot (not data class) of the component data type that is desired.

verbose

If the slot is empty, should a message be printed? values can be either 0 (no output) or 1 (print message)

Value

Returns the component object specified by the argument slot. Returns NULL if slot does not exist.

Examples

#
data(siamcat_example)
accessSlot(siamcat_example, "label")
#> $label
#>  CCIS27304052ST-3-0  CCIS15794887ST-4-0  CCIS74726977ST-3-0  CCIS16561622ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS79210440ST-3-0  CCIS82507866ST-3-0  CCIS98482370ST-3-0  CCIS00281083ST-3-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS32452666ST-4-0  CCIS94417875ST-3-0  CCIS51667829ST-4-0  CCIS27927933ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS45571137ST-3-0  CCIS81139242ST-4-0  CCIS21126322ST-4-0  CCIS17669415ST-4-0 
#>                  -1                  -1                  -1                   1 
#>  CCIS44676181ST-4-0  CCIS29210128ST-4-0  CCIS48725289ST-4-0 CCIS64785924ST-20-0 
#>                  -1                  -1                   1                  -1 
#>  CCIS03473770ST-4-0  CCIS10706551ST-3-0  CCIS64773582ST-4-0  CCIS06260551ST-3-0 
#>                  -1                  -1                  -1                   1 
#>  CCIS40244499ST-3-0  CCIS85214191ST-3-0  CCIS08668806ST-3-0 CCIS50369211ST-20-0 
#>                   1                   1                  -1                  -1 
#>  CCIS87167916ST-4-0  CCIS94496512ST-4-0  CCIS58234805ST-4-0  CCIS91228662ST-4-0 
#>                   1                  -1                   1                   1 
#>  CCIS71301801ST-4-0  CCIS24898163ST-4-0  CCIS53355328ST-4-0  CCIS41222843ST-4-0 
#>                   1                  -1                   1                   1 
#>  CCIS81887263ST-4-0 CCIS29688262ST-20-0  CCIS87116798ST-4-0  CCIS87605453ST-4-0 
#>                   1                  -1                   1                   1 
#>  CCIS83870198ST-4-0  CCIS61287323ST-4-0  CCIS11362406ST-4-0  CCIS14449628ST-4-0 
#>                   1                   1                  -1                   1 
#>  CCIS21278152ST-4-0  CCIS25399172ST-4-0  CCIS34055159ST-4-0 CCIS76845094ST-20-0 
#>                   1                  -1                   1                   1 
#>  CCIS41548810ST-4-0  CCIS98512455ST-4-0  CCIS83445808ST-4-0  CCIS37250421ST-4-0 
#>                   1                   1                  -1                  -1 
#>  CCIS35092938ST-4-0  CCIS50148151ST-4-0  CCIS62605362ST-3-0  CCIS87252800ST-4-0 
#>                  -1                   1                   1                   1 
#> CCIS22958137ST-20-0  CCIS48507077ST-4-0  CCIS32105356ST-4-0 CCIS81710917ST-20-0 
#>                   1                  -1                  -1                  -1 
#>  CCIS16326685ST-4-0  CCIS53043478ST-4-0  CCIS98832363ST-4-0  CCIS02124300ST-4-0 
#>                  -1                   1                   1                  -1 
#>  CCIS44743950ST-4-0  CCIS10793554ST-4-0  CCIS59132091ST-4-0  CCIS96387239ST-4-0 
#>                  -1                  -1                   1                  -1 
#>  CCIS82146115ST-4-0 CCIS31434951ST-20-0  CCIS33816588ST-4-0  CCIS12656533ST-4-0 
#>                   1                  -1                   1                   1 
#>  CCIS41806458ST-4-0  CCIS53557295ST-4-0  CCIS47284573ST-4-0  CCIS76563044ST-4-0 
#>                  -1                  -1                   1                  -1 
#> CCIS93040568ST-20-0  CCIS45793747ST-4-0  CCIS95097901ST-4-0  CCIS80834637ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS63910149ST-4-0  CCIS84543192ST-4-0 CCIS38765456ST-20-0 CCIS82944710ST-20-0 
#>                  -1                   1                   1                  -1 
#>  CCIS55230578ST-4-0  CCIS24254057ST-4-0  CCIS78318719ST-4-0  CCIS44757994ST-4-0 
#>                   1                   1                   1                   1 
#>  CCIS65479369ST-4-0  CCIS62794166ST-4-0  CCIS77252613ST-4-0  CCIS11015875ST-4-0 
#>                   1                   1                   1                   1 
#>  CCIS55531770ST-4-0  CCIS70398272ST-4-0  CCIS35100175ST-4-0  CCIS22416007ST-4-0 
#>                  -1                  -1                   1                   1 
#>  CCIS41288781ST-4-0  CCIS59903910ST-4-0  CCIS12370844ST-4-0  CCIS72607085ST-4-0 
#>                   1                  -1                   1                   1 
#>  CCIS46047672ST-4-0 CCIS22906510ST-20-0  CCIS52370277ST-4-0  CCIS02379307ST-4-0 
#>                   1                  -1                   1                   1 
#>  CCIS48579360ST-4-0  CCIS46467422ST-4-0  CCIS54027808ST-4-0  CCIS15704761ST-4-0 
#>                  -1                   1                  -1                   1 
#>  CCIS11558985ST-4-0  CCIS13047523ST-4-0  CCIS94603952ST-4-0  CCIS02856720ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS07648107ST-4-0  CCIS34604008ST-4-0  CCIS63468405ST-4-0  CCIS16383318ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS36699628ST-4-0  CCIS95409808ST-4-0  CCIS74239020ST-4-0  CCIS05314658ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS71578391ST-4-0  CCIS50003399ST-4-0  CCIS36797902ST-4-0  CCIS11354283ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS50471204ST-4-0  CCIS20795251ST-4-0  CCIS07277498ST-4-0  CCIS90164298ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS23164343ST-4-0  CCIS03857607ST-4-0  CCIS63448910ST-4-0  CCIS88007743ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS51595129ST-4-0  CCIS09568613ST-4-0  CCIS07539127ST-4-0  CCIS41692898ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS44093303ST-4-0  CCIS00146684ST-4-0  CCIS78100604ST-4-0  CCIS48174381ST-4-0 
#>                  -1                  -1                  -1                  -1 
#>  CCIS88317640ST-4-0 
#>                  -1 
#> 
#> $info
#> CTR CRC 
#>  -1   1 
#> 
#> $type
#> [1] "BINARY"
#> 
accessSlot(siamcat_example, "model_list")
#> model_list slot is empty.
#> NULL