Perform unsupervised feature filtering.

This function may convert absolute abundances into relative abundances and then performs unsupervised feature filtering. Features can be filtered based on abundance or prevalence. Additionally, unmapped reads may be removed.

filter.feat(feat, filter.method, cutoff, recomp.prop, rm.unmapped)

Arguments

feat
feature object

filter.method
method used for filtering the features, can be one of these: c("abundance", "cum.abundance", "prevalence")

cutoff
float, abundace or prevalence cutoff

recomp.prop
boolean, should absolute abundances be converted into relative abundances?

rm.unmapped
boolean, should unmapped reads be discarded?

Value

Returns the filtered feature matrix

Details

Currently, there are three filtering methods implemented:

"abundance" remove features for which the abundance is never above the threshold value (e.g. 0.5%) in any of the samples

"cum.abundance" remove features with very low abundance in all samples, i.e. features that are never among the most abundant entities that collectively make up (1-cutoff) of the reads in any samples

"prevalence" remove features with low prevalence across samples, i.e. features that are undetected in more than (1-cutoff) proportion of samples