Model Interpretation Plot

Produces a plot for model interpretation, displaying feature weights, robustness of feature weights, and features scores across patients.

interpretor.model.plot(siamcat, fn.plot, color.scheme = "BrBG",
  consens.thres = 0.5, heatmap.type = c("zscore", "fc"),
  norm.models = FALSE, limits = c(-3, 3), detect.lim = 1e-08,
  max.show = 50)

Arguments

siamcat
object of class siamcat-class

fn.plot
string, filename for the pdf-plot

color.scheme
color scheme for the heatmap, defaults to ="BrBG"

consens.thres
minimal ratio of models incorporating a feature in order to include it into the heatmap, defaults to 0.5 Note that for "randomForest" models, this cutoff specifies the minimum median Gini coefficient for a feature to be included and should therefore be much lower, e.g. 0.01

heatmap.type
type of the heatmap, can be either "fc" or "zscore", defaults to "zscore"

norm.models
boolean, should the feature weights be normalized across models?, defaults to FALSE

limits
vector, cutoff for extreme values in the heatmap, defaults to c(-3, 3)

detect.lim
float, pseudocount to be added before log10-transformation of features, defaults to 1e-08

max.show
integer, maximum number of features to be shown in the model interpretation plot, defaults to 50

Value

Does not return anything, but produces the model interpretion plot.

Details

Produces a plot consisting of

a barplot showing the feature weights and their robustness (i.e. in what proportion of models have they been incorporated)

a heatmap showing the z-scores of the metagenomic features across patients

another heatmap displaying the metadata categories (if applicable)

a boxplot displaying the poportion of weight per model that is actually shown for the features that are incorporated into more than consens.thres percent of the models.