Make predictions on a test set

This function takes a siamcat-class-object containing a model trained by train.model and performs predictions on a given test-set.

make.predictions(siamcat, siamcat.holdout = NULL,
    normalize.holdout = TRUE, verbose = 1)

Arguments

siamcat
object of class siamcat-class

siamcat.holdout
optional, object of class siamcat-class on which to make predictions, defaults to NULL

normalize.holdout
boolean, should the holdout features be normalized with a frozen normalization (see normalize.features) using the normalization parameters in siamcat?, defaults to TRUE

verbose
control output: 0 for no output at all, 1

for only information about progress and success, 2 for normal level of information and 3 for full debug information, defaults to 1

Value

object of class siamcat-class with the slot pred_matrix filled or a matrix containing the predictions for the holdout set

Details

This functions uses the model in the model_list-slot of the siamcat object to make predictions on a given test set. The test set can either consist of the test instances in the cross- validation, saved in the data_split-slot of the same siamcat object, or a completely external feature set, given in the form of another siamcat object (siamcat.holdout).

Examples

    data(siamcat_example)
    # Simple example
    siamcat.pred <- make.predictions(siamcat_example)

#> Error in label$label: $ operator not defined for this S4 class

    # Predictions on a holdout-set
    
# NOT RUN {
pred.mat <- make.predictions(siamcat.trained, siamcat.holdout,
    normalize.holdout=TRUE)
# }