Retrieve the information stored in the model_list slot within a SIAMCAT object

Function to retrieve the information stored in the model_list slot within a SIAMCAT object

model_list(siamcat, verbose=1)

# S4 method for siamcat
model_list(siamcat, verbose = 1)

Arguments

siamcat

(Required). An instance of siamcat-class that contains trained models

verbose

integer, if the slot is empty, should a message be printed? values can be either 0 (no output) or 1 (print message)

Value

The list stored in the model_list slot of the SIAMCAT object or NULL

Details

The function will return a list containing the information stored in the model_list slot of a SIAMCAT object. This list contains:

models - list of trained models

model_type - machine learning method used for training

feature_type - string describing on which type of features the models were trained

Examples

data(siamcat_example)
temp <- model_list(siamcat_example)
#> model_list slot is empty.
names(temp)
#> NULL