Perform feature normalization according to specified parameters.

This function performs feature normalization according to user-specified parameters.

normalize.feat(feat, norm.method = c("rank.unit", "rank.std", "log.std",
  "log.unit"), log.n0 = 10^-8, sd.min.q = 0.1, n.p = 2,
  n.sample = FALSE, n.feature = TRUE, n.global = FALSE)

Arguments

feat
feature object

norm.method
normalization method, can be one of these: c("rank.unit", "rank.std", "log.std", "log.unit")

log.n0
pseudocount to be added before log-transformation

sd.min.q
quantile of the distribution of standard deviation of all feature that will be added to the denominator during standardization of each feature in order to avoid underestimation (only for methods=c("rank.std", "log.std"))

n.p
vector norm to use, can be either 1 or 2, (only for method="log.unit")

n.sample
boolean, normalize by feature? (only for method="log.unit")

n.feature
boolean, normalize by sample? (only for method="log.unit")

n.global
boolean, Normalize by global rescaling? (only for method="log.unit")

Value

list containing the matrix of normalized features and a list of normalization parameters:

$par = parameters utilized in the normalization;

$feat = normalized features

Details

There are four different normalization methods implemented:

"rank.unit" converts features to ranks and normalizes each column by the square root of the sum of ranks

"rank.std" converts features to ranks and applies z-score standardization

"log.unit" log-transforms features (after addition of pseudocounts), converts features to ranks and normalizes by features or samples

"log.std" log-transforms features (after addition of pseudocounts) and applies z-score standardization

The other parameters are depending on the normalization method of choice, e.g. log.n0 as pseudocount is only needed for the methods making use of the log-transformation.