Model training

This function trains the a machine learning model on the training data, using a num.folds-fold internal cross-validation scheme to find the optimal hyper-parameters of the model.

plm.trainer(feat, label, cl = "classif.cvglmnet", data.split = NULL,
  stratify = TRUE, modsel.crit = "auc", min.nonzero.coeff = 1)

Arguments

feat
features object

label
label object

cl
class of learner, directly passed to makeLearner

data.split
filename containing the training samples or list of training instances produced by data.splitter(), defaults to NULL leading to training on the complete dataset

stratify
boolean, should the folds in the internal cross-validation be stratified?

min.nonzero.coeff
integer number of minimum nonzero coefficients that should be present in the model

Value

an object of class makeWrappedModel