Retrieve the prediction matrix from a SIAMCAT object

Function to retrieve the prediction matrix from a SIAMCAT object

pred_matrix(siamcat, verbose=1)

# S4 method for siamcat
pred_matrix(siamcat, verbose = 1)

Arguments

siamcat

(Required). A siamcat-class object that contains a prediction matrix

verbose

integer, if the slot is empty, should a message be printed? values can be either 0 (no output) or 1 (print message)

Value

A matrix containing predictions or NULL

Details

The functions returns a matrix containing the predictions for all samples across the different cross-validation repeats. See make.predictions for more information.

Examples

data(siamcat_example)
temp <- pred_matrix(siamcat_example)
head(temp)
#>                        CV_rep1     CV_rep2
#> CCIS27304052ST-3-0 0.024907873 0.099056615
#> CCIS15794887ST-4-0 0.008698458 0.013123270
#> CCIS74726977ST-3-0 0.802378196 0.996027280
#> CCIS16561622ST-4-0 0.066810882 0.009024059
#> CCIS79210440ST-3-0 0.099713743 0.233315321
#> CCIS82507866ST-3-0 0.022160415 0.262949108