Select samples based on metadata

This functions filters features, labels, and metadata based on a specific column in the metadata. Provided with a column in the metadata and a range or a set of allowed values, the function will return the labels, metadata, and features for the samples matching the allowed range or set.

select.samples(meta, feat, label, filter, allowed.set = NULL,
  allowed.range = NULL)

Arguments

meta
metadata object

feat
features object

label
labels object

filter
name of the column from metadata on which the selection should be done

allowed.set
a vector of allowed values

allowed.range
a range of allowed values

Value

list containing values after selection:

$feat = features;

$label = labels;

$meta = metadata