The S4 SIAMCAT class

The SIAMCAT class

Details

The S4 SIAMCAT class stores the results from the SIAMCAT workflow in different slots. The different slots will be filled by different functions (referenced in the description below).

In order to contruct a SIAMCAT class object, please refer to the documentation of the construction function siamcat.

The SIAMCAT class is based on the phyloseq-class. Therefore, you can easily import a phyloseq object into SIAMCAT.

Slots

phyloseq

object of class phyloseq-class

label

list containing the label information for the samples and some metadata about the label, created by create.label or when creating the siamcat-class object by calling siamcat

filt_feat

list containing the filtered features as matrix and the list of filtering parameters, created by calling the filter.features function

associations

list containing the parameters for association testing and the results of association testing with these parameters in a dataframe, created by calling the check.associations function

norm_feat

list containing the normalized features as matrix and the list of normalziation parameters (for frozen normalization), created by calling the normalize.features function

data_split

list containing cross-validation instances, created by calling the create.data.split function

model_list

list containing the trained models, the type of model that was trained, and on which kind of features it was trained, created by calling the train.model function

pred_matrix

matrix of predictions, created by calling the make.predictions function

eval_data

list containing different evaluation metrics, created by calling the evaluate.predictions function