Add metadata as predictors

This function adds metadata to the feature matrix to be later used as predictors

add.meta.pred(siamcat, pred.names = NULL, std.meta = TRUE, verbose = 1)

Arguments

siamcat
object of class siamcat-class

pred.names
vector of names of the variables within the metadata to be added to the feature matrix as predictors

std.meta
boolean, should added metadata features be standardized?, defaults to TRUE

verbose
control output: 0 for no output at all, 1 for only information about progress and success, 2 for normal level of information and 3 for full debug information, defaults to 1

Value

an object of class siamcat-class with metadata added to the features

Examples

 data(siamcat_example)
 # Add the Age of the patients as potential predictor
 siamcat_age_added <- add.meta.pred(siamcat_example, pred.names=c('age'))

#> Adding metadata as predictor finished

 # Add Age, BMI, and Gender as potential predictors
 # Additionally, prevent standardization of the added features
 siamcat_meta_added <- add.meta.pred(siamcat_example, pred.names=c('age',
     'bmi', 'gender'), std.meta=FALSE)

#> Adding metadata as predictor finished