Check and visualize associations between features and classes

This function calculates for each feature the median log10 fold change between the different classes found in labels.

Significance of the differences is computed for each feature using a wilcoxon test followed by multiple hypothesis testing correction.

Additionally, the Area under the Receiver Operating Characteristic Curve (AU-ROC) is computed for the features found to be associated with the two different classes at a user-specified significance level alpha.

Finally, the function produces a plot of the top max.show associated features, showing the distribution of the log10-transformed abundances for both classes, the median fold change, and the AU-ROC.

check.associations(feat, label, fn.plot, color.scheme = "RdYlBu",
  alpha = 0.05, min.fc = 0, mult.corr = "fdr", detect.lim = 10^-8,
  max.show = 50, plot.type = "bean")

Arguments

feat
feature object

label
label object

fn.plot
filename for the pdf-plot

color.scheme
valid R color scheme, defaults to 'RdYlBu'

alpha
float, significance level, defaults to 0.05

min.fc
float, minimum log10 fold change for significantly associated features, defaults to 0

mult.corr
multiple hypothesis correction method, see p.adjust, defaults to "fdr"

detect.lim
float, pseudocount to be added before log-transormation of the data, defaults to 1e-08

max.show
integer, how many associated features should be shown, defaults to 50

plot.type
string, specify how the abundance should be plotted, must be one of these: c("bean", "box", "quantile.box", "quantile.rect"), defaults to "bean"

Value

Does not return anything, but produces an association plot