Model Evaluation Plot

Produces two plots for model evaluation. The first plot shows the Receiver Operating Characteristic (ROC)-curves, the other the Precision-recall (PR)-curves for the different cross-validation repetitions.

model.evaluation.plot(..., fn.plot = NULL,
    colours=NULL, show.all=FALSE, verbose = 1)

Arguments

...
one or more object of class siamcat-class, can be named

fn.plot
string, filename for the pdf-plot

colours
colour specification for the different siamcat-class- objects, defaults to NULL which will cause the colours to be picked from the 'Set1' palette

show.all
boolean, Should all repeated cross-validation models be plotted?

verbose
control output: 0 for no output at all, 1 for only information about progress and success, 2 for normal level of information and 3 for full debug information, defaults to 1

Value

Does not return anything, but produces the model evaluation plot.

Examples

data(siamcat_example)

# simple working example
model.evaluation.plot(siamcat_example, fn.plot='./eval.pdf')

#> Plotted evaluation of predictions successfully to: ./eval.pdf

# plot several named SIAMCAT object
# (although we use only one example object here)
model.evaluation.plot('Example_1'=siamcat_example,
    'Example_2'=siamcat_example, colours=c('red', 'blue'),
    fn.plot='./eval.pdf')

#> Plotted evaluation of predictions successfully to: ./eval.pdf