Summarize features

This function summarize features on a specific taxonomic level

summarize.features(siamcat, level = 'g__',
                    feature.type='original', verbose=1)

Arguments

siamcat
object of class siamcat-class

level
string, at which level to summarize (e.g. g__ = genus)

feature.type
string, on which type of features should the function work? Can be either "original", "filtered", or "normalized". Please only change this paramter if you know what you are doing!

verbose
integer, control output: 0 for no output at all, 1 for only information about progress and success, 2 for normal level of information and 3 for full debug information, defaults to 1

Value

object of class siamcat-class with a summarized feature table

Details

This function will summarize features at different taxonomic levels, e.g. transform species-level relative abundance into genus-level taxonomic profiles.

The function expects a SIAMCAT object that either contains an entry in the phyloseqtax_table slot of its phyloseq object, OR a set of feature names which encode taxonomic information, e.g.

k__Bacteria;p__Actinobacteria;c__Actinobacteria;o__Acidimicrobiales;..

Then, for a given taxonomic level (e.g. g__), the function will sum up all the relative abundances of features belonging to the same group at that specific taxonomic level.

Please note that this function is currently maturing and not necessarily reliable!!!

Examples

## load the phyloseq example data
data("GlobalPatterns")
## create an example label
label <- create.label(meta=sample_data(GlobalPatterns),
    label = "SampleType",
    case = c("Freshwater", "Freshwater (creek)", "Ocean"))

#> Label used as case:
#>    Freshwater,Freshwater (creek),Ocean
#> Label used as control:
#>    rest

#> + finished create.label.from.metadata in 0.002 s

# run the constructor function
siamcat <- siamcat(phyloseq=GlobalPatterns, label=label, verbose=1)

#> Warning: ### Warning: The data do not seem to consist of relative abundances! (values ranging between 0 and 1)

#> Data set has only 8 training examples of class Case .
#> Note that a dataset this small/skewed is not necessarily suitable for analysis in this pipeline.

#> Data succesfully validated

siamcat <- summarize.features(siamcat, level='Genus', verbose=3)

#> + starting summarize.features

#> +++ summarizing on level: Genus

#> +++ summarized features table contains: 984 features

#> Warning: Tax table does not seem to be consistent in all cases...
#> Will be collapsed at level Genus

#> Warning: Phylogenetic tree in original SIAMCAT object had to be deleted...

#> + finished summarize.features in 0.706 s