Model training

This function trains the a machine learning model on the training data

train.model(siamcat, method = c("lasso", "enet", "ridge", "lasso_ll",
  "ridge_ll", "randomForest"), data.split = NULL, stratify = TRUE,
  modsel.crit = list("auc"), min.nonzero.coeff = 1, param.set = NULL)

Arguments

siamcat
object of class siamcat-class

method
string, specifies the type of model to be trained, may be one of these: c("lasso", "enet", "ridge", "lasso_ll", "ridge_ll", "randomForest")

data.split
filename containing the training samples or list of training instances produced by data.splitter(), defaults to NULL leading to training on the complete dataset

stratify
boolean, should the folds in the internal cross-validation be stratified?

modsel.crit
list, specifies the model selection criterion during internal cross-validation, may contain these: c("auc", "f1", "acc", "pr")

min.nonzero.coeff
integer number of minimum nonzero coefficients that should be present in the model (only for "lasso", "ridge", and "enet"

param.set
a list of extra parameters for mlr run, may contain: cost - for lasso_ll and ridge_ll; alpha for enet and ntree, mtry for RandomForrest

Value

object of class siamcat-class